refactor(server-cloudflare): consolidate auth into single src/auth.ts singleton
Inline options.ts and remove createAuth() wrapper — the auth singleton
is now a single top-level betterAuth() call in src/auth.ts. The CLI
config (better-auth.config.ts) duplicates schema-affecting options
since it runs in Bun, not Workers.
