+++ packages/server-cloudflare/better-auth.config.ts
@@ -8,20 +8,20 @@
  *
  * These scripts load `.dev.vars` automatically via `--env-file`.
  *
- * This file spreads {@link authOptions} (from `src/auth/options.ts`) so the
- * CLI schema always matches the runtime worker. Runtime-only options (KV
- * caching, cookies, trusted origins) are omitted — they don't affect the
- * database schema.
+ * Schema-affecting options (basePath, plugins, emailAndPassword) are duplicated
+ * from `src/auth.ts` — keep them in sync. Runtime-only options (KV caching,
+ * cookies, trusted origins) are omitted since they don't affect the schema.
  *
- * @see src/auth/server.ts — runtime `createAuth()` factory
- * @see src/auth/options.ts — shared schema-affecting options
+ * @see src/auth.ts — runtime singleton (source of truth)
  */
 import { type } from 'arktype';
 import { neon } from '@neondatabase/serverless';
 import { betterAuth } from 'better-auth';
 import { drizzleAdapter } from 'better-auth/adapters/drizzle';
+import { bearer } from 'better-auth/plugins/bearer';
+import { jwt } from 'better-auth/plugins/jwt';
+import { oauthProvider } from '@better-auth/oauth-provider';
 import { drizzle } from 'drizzle-orm/neon-http';
-import { authOptions } from './src/auth/options';
 
 const CliEnv = type({
 	DATABASE_URL: 'string',
@@ -38,8 +38,38 @@ if (env instanceof type.errors) {
 const sql = neon(env.DATABASE_URL);
 const db = drizzle(sql);
 
+// Keep schema-affecting options in sync with src/auth.ts
 export const auth = betterAuth({
-	...authOptions,
+	basePath: '/auth',
+	emailAndPassword: { enabled: true },
+	plugins: [
+		bearer(),
+		jwt(),
+		oauthProvider({
+			loginPage: '/sign-in',
+			consentPage: '/consent',
+			requirePKCE: true,
+			allowDynamicClientRegistration: true,
+			trustedClients: [
+				{
+					clientId: 'epicenter-desktop',
+					name: 'Epicenter Desktop',
+					type: 'native',
+					redirectUrls: ['tauri://localhost/auth/callback'],
+					skipConsent: true,
+					metadata: {},
+				},
+				{
+					clientId: 'epicenter-mobile',
+					name: 'Epicenter Mobile',
+					type: 'native',
+					redirectUrls: ['epicenter://auth/callback'],
+					skipConsent: true,
+					metadata: {},
+				},
+			],
+		}),
+	],
 	database: drizzleAdapter(db, { provider: 'pg' }),
 	secret: env.BETTER_AUTH_SECRET,
 });
